# docs/openapi/pn-safestorage-internal-v1-api.yaml
openapi: 3.0.3
info:
  title: >-
    SafeStorageInternal
  version: v1.0
  license:
    name: PN software license
    url: 'https://www.pn.pagopa.it/LICENSE'
  description: >-
    __SafeStorageInternal__: CRUD interne
servers:
  - url: >-
      https://api.pn.pagopa.it
    description: Server url
...
paths:
...
  '/safestorage/internal/v1/userConfigurations/{name}':
    get:
      operationId: getUserConfiguration
      summary: Restituisce una anagrafica client.
      tags:
        - userConfigurationInternal
      parameters:
        - $ref: '#/components/parameters/name'
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/userConfigurations/TEST_CLIENT
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
              example:
                docType: null
                error:
                  code: null
                  description: Client id CLIENT_TEST not found
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConfigurationResponse'
    patch:
      operationId: patchUserConfiguration
      summary: Aggiorna una anagrafica client.
      tags:
        - userConfigurationInternal
      parameters:
        - $ref: '#/components/parameters/name'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserConfigurationChanges'
        required: true
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/userConfigurations/TEST_CLIENT
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
              example:
                docType: null
                error:
                  code: null
                  description: Client id CLIENT_TEST not found
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConfigurationResponse'
    delete:
      operationId: deleteUserConfiguration
      summary: Elimina una anagrafica client.
      tags:
        - userConfigurationInternal
      parameters:
        - $ref: '#/components/parameters/name'
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/userConfigurations/TEST_CLIENT
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
              example:
                docType: null
                error:
                  code: null
                  description: Client id CLIENT_TEST not found
        '204':
          description: OK
  '/safestorage/internal/v1/userConfigurations':
    post:
      operationId: insertUserConfiguration
      summary: Inserisce una nuova anagrafica client.
      tags:
        - userConfigurationInternal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserConfiguration'
      responses:
        '400':
          description: >-
            #### Bad request
        '401':
          description: >-
            #### Unauthorized
              - __x-api-key__ e/o __x-pagopa-safestorage-cx-id__ assenti
        '403':
          description: >-
            #### Forbidden
              - Invalid __x-api-key__
          content:
            application/json:
              example:
                timestamp: 2023-02-21T10:24:06.294+00:00
                path: /safestorage/internal/v1/userConfigurations/TEST_CLIENT
                status: 403
                error: Forbidden
                message: Invalid x-api-key
                requestId: cdf95798-11
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserConfigurationResponse'
components:
  parameters:
...
    name:
      name: name
      description: Nome del client
      in: path
      required: true
      schema:
        type: string
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    DocumentType:
      type: object
      properties:
        tipoDocumento:
          type: string
          description: nome del tipo di documento
        checksum:
          description: >-
            codice di verifica del corretto caricamento del file
            https://aws.amazon.com/it/blogs/aws/new-additional-checksum-algorithms-for-amazon-s3/
          type: string
          enum: [ 'MD5', 'SHA256', 'NONE']
        initialStatus:
          type: string
        statuses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CurrentStatus'
        informationClassification:
          description: >-
            * `C` - Confidential
            * `HC`- Highly Confidential </br>
            Se il valore è HC il file deve essere criptato.</br>
            See [information classification](https://pagopa.atlassian.net/wiki/spaces/EN/pages/357204284/Data+Classification+Handling#Classificazione)
          type: string
          enum: [ 'C', 'HC' ]
        transformations:
          type: array
          items:
            type: string
            enum: ['NONE', 'SIGN_AND_TIMEMARK']
        timeStamped:
          type: string
          enum: [ 'NONE', 'STANDARD' ]
    CurrentStatus:
      type: object
      properties:
        allowedStatusTransitions:
          $ref: '#/components/schemas/AllowedStatusTransitions'
        storage:
          $ref: '#/components/schemas/Storage'
        technicalState:
          $ref: '#/components/schemas/TechnicalState'
    Storage:
      type: string
    AllowedStatusTransitions:
      type: array
      items:
        type: string
    TechnicalState:
      type: string
...
    Document:
      type: object
      properties:
        documentKey:
          type: string
        documentState:
          type: string
        documentLogicalState:
          type: string
        clientShortCode:
          type: string
        retentionUntil:
          description: >-
            data di scadenza del file
          type: string
        checkSum:
          description:
            valore ottenuto dal calcolo della chiave hashing dell’oggetto
          type: string
        contentLenght:
          type: number
        contentType:
          type: string
        documentType:
          $ref: '#/components/schemas/DocumentType'
        lastStatusChangeTimestamp:
          type: string
          format: date-time
          description: >-
            Timestamp dell'ultima variazione di stato del documento.
...
    DocumentResponse:
      type: object
      properties:
        document:
          allOf:
            - $ref: '#/components/schemas/Document'
            - type: object
              properties:
                documentType:
                  $ref: '#/components/schemas/DocumentType'
        error:
          $ref: '#/components/schemas/Error'
...
    UserConfiguration:
      type: object
      properties:
        name:
          type: string
        apiKey:
          type: string
        canCreate:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a creare.
        canRead:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a leggere.
        canModifyStatus:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a modificare.
        signatureInfo:
          type: string
        destination:
          type: object
          description: Informazioni per contattare in modalità push i sistemi client.
          properties:
            sqsUrl:
              type: string
              description: URL della coda SQS a cui contattare il client
    UserConfigurationResponse:
      type: object
      properties:
        userConfiguration:
          $ref: '#/components/schemas/UserConfiguration'
        error:
          $ref: '#/components/schemas/Error'
    UserConfigurationChanges:
      type: object
      properties:
        canCreate:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a creare.
        canRead:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a leggere.
        canModifyStatus:
          type: array
          items:
            type: string
          description: Elenco dei tipi di documenti che il client è autorizzato a modificare.
        apiKey:
          type: string
        signatureInfo:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: [] # use the same name as under securitySchemes